Fix no-default-features build
