+++ src/model/data_availability.rs
@@ -2,7 +2,9 @@ use bincode::{Decode, Encode};
 use std::collections::HashMap;
 
 use crate::model::{BlobsHash, BlockHeight, ContractName};
-use hyle_contract_sdk::{Blob, HyleOutput, Identity, ProgramId, StateDigest, TxHash, Verifier};
+use hyle_contract_sdk::{
+    Blob, BlobIndex, HyleOutput, Identity, ProgramId, StateDigest, TxHash, Verifier,
+};
 use serde::{Deserialize, Serialize};
 
 #[derive(Default, Debug, Clone, Serialize, Deserialize, Encode, Decode)]
@@ -26,6 +28,18 @@ pub struct UnsettledBlobMetadata {
     pub blob: Blob,
     // Each time we receive a proof, we add it to this list
     pub possible_proofs: Vec<(ProgramId, HyleOutput)>,
+}
+
+#[derive(
+    Debug, Default, Clone, serde::Serialize, serde::Deserialize, Encode, Decode, Eq, PartialEq,
+)]
+pub struct HandledBlobProofOutput {
+    pub proof_tx_hash: TxHash,
+    pub blob_tx_hash: TxHash,
+    pub blob_index: BlobIndex,
+    pub contract_name: ContractName,
+    pub hyle_output: HyleOutput,
+    pub blob_proof_output_index: usize,
 }
 
 #[derive(Default, Debug, Clone, Encode, Decode)]
